+++ activerecord/lib/active_record/migration.rb
@@ -574,6 +574,7 @@ module ActiveRecord
     autoload :JoinTable, "active_record/migration/join_table"
     autoload :ExecutionStrategy, "active_record/migration/execution_strategy"
     autoload :DefaultStrategy, "active_record/migration/default_strategy"
+    autoload :SqlFilesystemVersions, "active_record/migration/sql_filesystem_versions"
 
     # This must be defined before the inherited hook, below
     class Current < Migration # :nodoc: